Update ActorPrevalence.sql
Revamped version of the SQL code, including new comments.
This version creates a table that counts how many actors have certain movie/show combinations. That metric is important for my data viz because it will most intuitively show if there is a bias to be movie/show actors or a combination.